회원가입 기능 구현(test 미실시)

+++ src/main/java/kr/co/easystock/domain/user/User.java
@@ -17,11 +17,10 @@ import java.time.LocalDateTime;
 @Table
 public class User extends BaseTimeEntity
 {
-
     @Id
     @Column
     @GeneratedValue(strategy = GenerationType.IDENTITY) // auto increment
-    private Long id;
+    private int id;
 
     @Column
     private String name;
